refactor(notification): extract FCM event parsers into notification_event_parser
Move the pure base64/protobuf push-decoding logic (extract_notification_id,
_extract_event_with_compiled_protos, _resolve_qualifier, _find_embedded_messages,
_extract_source_info, _extract_space_source_info, _extract_space_group_info,
_extract_event_raw) plus the group-id sanity constants out of notification.py
into a new side-effect-free notification_event_parser.py.

AjaxNotificationListener keeps thin delegators for the methods that have
callers (extract_notification_id, _extract_event_with_compiled_protos kept as
an instance method, _extract_source_info, _extract_space_source_info,
_extract_space_group_info, _extract_event_raw); _parse_and_fire_event still
orchestrates them and owns the side effects (firing HA events, applying
security state). The parser module pins its logger name to notification so
DEBUG output and the caplog-based tests keep their namespace.

The four white-box priority-resolution tests mocked the _find_embedded_messages
/ _resolve_qualifier seam via patch.object(listener, ...); that seam moved to
the module, so they now patch notification_event_parser.* — same behaviour
asserted, relocated target. Every other test is unchanged.

notification.py drops from 943 to ~590 lines. Pure refactor — no behaviour
change, 1396 tests pass, 88% coverage. Tier-2 #1 from the refactor backlog.
